pyproject.toml
[project]
name = "graph-code"
version = "0.0.24"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "mcp>=1.21.1",
    "pydantic-ai-slim[google,openai,vertexai]>=0.2.18",
    "pydantic-settings>=2.0.0",
    "pymgclient>=1.4.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.0",
    "toml>=0.10.2",
    "tree-sitter-python>=0.23.6",
    "tree-sitter==0.25.0",
    "watchdog>=6.0.0",
    "typer>=0.12.5",
    "rich>=13.7.1",
    "prompt-toolkit>=3.0.0",
    "diff-match-patch>=20241021",
    "click>=8.0.0",
    "pytest-xdist>=3.8.0",
    "pydantic-ai[anthropic]>=1.28.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
]

[project.scripts]
graph-code = "codebase_rag.main:app"

[tool.uv]
package = true

[tool.setuptools]
packages = ["codebase_rag"]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
dev = [
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "pre-commit>=4.2.0",
]

treesitter-full = [
    "tree-sitter-python>=0.23.6",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-scala>=0.24.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-lua>=0.0.19",
]


[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["codebase_rag/tests"]
norecursedirs = ["grammars", ".git", "__pycache__", "*.egg-info", ".venv", "venv"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pyinstaller>=6.14.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-toml>=0.10.8.20240310",
]
